requirements.txt
python-telegram-bot==13.15
python-dotenv==1.1.1
aiogram==3.21.0